Turn on warnings for TODO and FIXME comments
TODO and FIXME comments are dangerous and can be forgotten about. It is
better to keep things like that in project/task management software so
it is not forgotten about; however, that can sometimes be difficult
depending on the project management system used. As a simple default, we
can enable warnings for these so they are seen every time the project is
linted.